pd: 🔐 extract auto-https into a standalone crate
fixes #3119. see also, #1886.

this pulls the auto-https code (see #3627, #3652) into a standalone
library crate.
