+++ crates/bin/pd/src/lib.rs
@@ -11,7 +11,6 @@ mod mempool;
 mod metrics;
 mod snapshot;
 
-pub mod auto_https;
 pub mod cli;
 pub mod events;
 pub mod migrate;
