+++ crates/util/auto-https/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "penumbra-auto-https"
+version = "0.65.0-alpha.1"
+authors = ["Penumbra Labs <team@penumbra.zone>"]
+edition = "2021"
+description = "Automatic HTTPS management for Penumbra"
+repository = "https://github.com/penumbra-zone/penumbra/"
+homepage = "https://penumbra.zone"
+license = "MIT OR Apache-2.0"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+futures = "0.3"
+rustls = "0.20.9"
+axum-server = { version = "0.4.7", features = [] }
+rustls-acme = { version = "0.6.0", features = ["axum"] }
+tracing = "0.1"
